+++ src/components/settings/FilesSorting.vue
@@ -11,7 +11,9 @@
       </div>
       <div class="control is-expanded">
         <div class="select is-fullwidth">
-          <select v-model="localSortingDirection">
+          <select
+            :class="{ 'is-lowercase': !useNormalLetterCase }"
+            v-model="localSortingDirection">
             <option value="default">Default (based on the field)</option>
             <option value="asc">Ascending</option>
             <option value="desc">Descending</option>
@@ -37,7 +39,7 @@
         id="sorting-files-size"
         value="size"
         v-model="localSorting">
-      <label for="sorting-files-size">Sort by filesize</label>
+      <label for="sorting-files-size">Sort by file size</label>
     </div>
 
     <div class="field">
@@ -128,6 +130,7 @@
         <button
           type="button"
           class="button is-primary"
+          :class="{ 'is-lowercase': !useNormalLetterCase }"
           @click="addNamespace">
           Add
         </button>
@@ -141,6 +144,8 @@
 <script>
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome'
 import Draggable from 'vuedraggable'
+
+import config from '@/config'
 
 export default {
   name: 'FilesSorting',
@@ -160,7 +165,8 @@ export default {
   },
   data: function () {
     return {
-      newNamespace: ''
+      newNamespace: '',
+      useNormalLetterCase: config.useNormalLetterCase
     }
   },
   computed: {