+++ src/config/index.js
@@ -24,13 +24,14 @@ export default {
   version: '2.2.1',
   title: process.env.VUE_APP_TITLE,
   subtitle: process.env.VUE_APP_SUBTITLE,
+  useNormalLetterCase: process.env.VUE_APP_USE_NORMAL_LETTER_CASE === 'true',
   apiUrl: process.env.VUE_APP_API_URL,
-  registrationIsEnabled: (process.env.VUE_APP_REGISTRATION_ENABLED === 'true'),
+  registrationIsEnabled: process.env.VUE_APP_REGISTRATION_ENABLED === 'true',
   authenticationIsRequired: process.env.VUE_APP_AUTHENTICATION_REQUIRED
-    ? (process.env.VUE_APP_AUTHENTICATION_REQUIRED === 'true')
+    ? process.env.VUE_APP_AUTHENTICATION_REQUIRED === 'true'
     : true,
   minPasswordLength: process.env.VUE_APP_MIN_PASSWORD_LENGTH || 16,
-  countsAreEnabled: (process.env.VUE_APP_COUNTS_ENABLED === 'true'),
+  countsAreEnabled: process.env.VUE_APP_COUNTS_ENABLED === 'true',
   fallbackFilesSortingNamespace:
     process.env.VUE_APP_FALLBACK_FILES_SORTING_NAMESPACE || 'namespace',
   defaultNamespaceColors: generateNamespaceColors(